Create method `with()` on `ResultInterface`
It doesn't make sense to allow a person to change a result, it must be
immutable but some rules, like `not()`, `key()` and the `Validator`
class itself, need to change the result. The `not()` rule, for example,
needs to change not only the status (`isValid()`) but also change the
mode of the message to be shown by the exception.

For the reasons above the method `with()` is necessary, then we can
create new results changing only what is necessary and keeping the
result immutable.